[Ingest] Some entity matching naming / cleanup (Recidiviz/recidiviz-data#17539)
## Description of the change

Some entity matching code cleanup to prepare for a refactor to support
non-StatePerson root entities.

This PR:
* Deletes all code gated by `do_ingest_obj_consistency_check` and
`log_entity_counts` which are always true (not sure when they were used
before)
* Renames `BaseStateMatchingDelegate` to
`StateSpecificEntityMatchingDelegate` to indicate this delegate should
house state-specific matching logic. There will likely be a new delegate
that is introduced for abstracting away logic that is specific to the
root entity type (such as reading that entity from the DB).
* Ensures no instance variable names in the `StateEntityMatcher
`reference `person` or `people`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#17471.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 563f09fb1bd4dc4d3e816cb790b433a80fd3a2eb