[Workflows] Remove state_code prefix from specified criteria (Recidiviz/recidiviz-data#17563)
## Description of the change

The common criterion for SLD across states is of the form
`US_XX_SUPERVISION_LEVEL_HIGHER_THAN_ASSESSMENT_LEVEL`. To make life
easier on the frontend, we want to remove `US_XX_` from these names in
the ETL process.

This PR creates a list of criteria for which we should remove the
initial state code prefix. Right now, it only includes
`SUPERVISION_LEVEL_HIGHER_THAN_ASSESSMENT_LEVEL`.

To do this, it was worth constructing a set of known names to look for
to reuse on every step instead of running a regex on every single
criterion name. However, this required making `state_code` a class
property. The refactor to do this is in the first commit and the actual
work to remove the prefix is in the second commit.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#17552

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 73d2ef616363cdb0c9ad1ce11b74a08e7cbad11a