[Justice Counts] send email address with agency information (Recidiviz/recidiviz-data#17658)
## Description of the change

This PR updates the `to_json` method for agencies to include `email`
with user information. This will be used in the Team Management section
on the Agency Settings page.

## Related issues

Closes to Recidiviz/recidiviz-data#17645

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 218c1be9f85718b4f74d63dffc11593550951960