[Justice Counts] Performance improvements (Recidiviz/recidiviz-data#17398)
## Description of the change

A few small tweaks that may slightly help performance:
* I’m moving the `raise_if_user_is_unauthorized` check into the body of
the endpoint, rather than having it as a decorator. This allows us to
eliminate an unnecessary call to the database.
* I’m no longer storing the UserAccount object on the session context.
To do this, we had to make a call to the database in our
`on_successful_authorization` decorator. Now we can remove that call to
the database and just fetch the UserAccount object later if we really
need it.
* Also getting rid of some logging that we weren’t using and were just
spamming our logs. This won't affect perf but was just bugging me while
I was investigating perf :)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3ac060aa2beef02fbe0905f3e89c6ac99e1d6d4b