[US_TN] Updating sentence_metadata and completion_date fields  (Recidiviz/recidiviz-data#17689)
## Description of the change

After the rerun flash on 1/5/23, Damini noted some behaviors that caused
issues downstream with the Sentences v2 view in TN that includes ISC and
Diversion sentences.

This PR adds in the sentence_source to sentence_metadata to avoid having
to parse the external_id at any point due to certain ISC sentences have
"-" in their CaseNumbers.

It also uses mappings to null out completion_dates of "9999-12-31" to
avoid overflow errors.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8ac708e780efd15594dc4853a741c2d2212f4bf2