[Ingest] Restructure Entity superclasses to allow for more generic entity matching logic (Recidiviz/recidiviz-data#17669)
## Description of the change

Establishes some useful Entity superclasses that will allow us to deal
with `StatePerson`/`StateStaff` and
`StatePersonExternalId`/`StateStaffExternalId` more generically.

In this PR:
* Renames `ExternalIdEntity` -> `HasExternalIdEntity`
* Adds new `ExternalIdEntity` superclass, currently used by
`StatePersonExternalId` but soon to be used by `StateStaffExternalId`
* Adds new `HasMultipleExternaIdsEntity` superclass, currently used by
`StatePerson` but soon to be used by `StateStaff`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#17471.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2165243ac3636f7efb0dfc032ec0c9c08dfabe45