[Justice Counts] add role and invitation status to agency_user_account_association_table (Recidiviz/recidiviz-data#17719)
## Description of the change

This is a prerequisite for the team management work stream. The role /
invitation of a user will depend on what agency they are in. For that
reason, we cannot store them in the user record and have to store them
on the agency_user_account_association_table.

## Related issues

Related to Recidiviz/recidiviz-data#17382

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f23a826a0533501ebfc00167ef116253e346a7f0