[Justice Counts] arrests metric (Recidiviz/recidiviz-data#17650)
## Description of the change

This updates the arrests metric for law enforcement.

I plan on following up with a refactor of the `OffenseType` dimension so
that, depending on the metric, you get the right label. For example, the
label for the breakdowns _should_ end with Charges/Offenses (i.e Drug
Charges / Offenses). It didn't seem right to create a whole new
dimension to capture this change.

## Related issues

Closes Recidiviz/recidiviz-data#17241
Contributes to Recidiviz/recidiviz-data#17522

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d6553a8403d1465deb970b57e4056bbacd139440