[US_MO] Update housing stay stitching logic (Recidiviz/recidiviz-data#17368)
## Description of the change

Missouri uses a handful of codes to describe the "type" of housing
someone is in: general population, AdSeg, and temporary AdSeg, among
others; and also whether that housing is "permanent" or "temporary".
This PR updates the logic to stitch stays together of the same "type" to
more closely fit the way solitary confinement policies are understood in
the MO DOC in a couple ways:
1. "TEMPORARY" is broken into "TEMPORARY-TASC" (for all stays of type
"TAS") and "TEMPORARY-OTHER" (for all "temporary" stays of any other
type).
2. The codes `DIS` (disciplinary segregation) and `NOC` ("Ad Seg No
Contact") are considered solitary confinement (rather than unknown).
`PRC` (protective custody) and `HOS` (hospital) are given their own
confinement types for sessions views.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#15359

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d42e42dff985081cc00dc0de9e4802c14f3d68ed